:recycle: move answerStatus to share dir
