+++ api/lib/domain/services/solution-service-qcm.js
@@ -1,5 +1,5 @@
 import { _ } from '../../infrastructure/utils/lodash-utils.js';
-import { AnswerStatus } from '../../../src/school/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../src/shared/domain/models/AnswerStatus.js';
 
 const match = function (answer, solution) {
   if (_.areCSVequivalent(answer, solution)) {
