+++ api/lib/domain/services/solution-service-qroc.js
@@ -6,7 +6,7 @@ import {
 import lodash from 'lodash';
 import { applyTreatments, applyPreTreatments } from './validation-treatments.js';
 import { validateAnswer } from './string-comparison-service.js';
-import { AnswerStatus } from '../../../src/school/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../src/shared/domain/models/AnswerStatus.js';
 import { getEnabledTreatments, useLevenshteinRatio } from './services-utils.js';
 const CHALLENGE_NUMBER_FORMAT = 'nombre';
 const { every, isEmpty, isString, map } = lodash;
