+++ api/src/evaluation/infrastructure/repositories/answer-repository.js
@@ -3,7 +3,7 @@ import jsYaml from 'js-yaml';
 import { knex } from '../../../../db/knex-database-connection.js';
 import { ChallengeAlreadyAnsweredError, NotFoundError } from '../../../../lib/domain/errors.js';
 import { Answer } from '../../domain/models/Answer.js';
-import * as answerStatusDatabaseAdapter from '../../../../src/school/infrastructure/adapters/answer-status-database-adapter.js';
+import * as answerStatusDatabaseAdapter from '../../../../src/shared/infrastructure/adapters/answer-status-database-adapter.js';
 
 function _adaptAnswerToDb(answer) {
   return {
