+++ api/src/school/infrastructure/repositories/activity-answer-repository.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import jsYaml from 'js-yaml';
 import { knex } from '../../../../db/knex-database-connection.js';
-import * as answerStatusDatabaseAdapter from '../adapters/answer-status-database-adapter.js';
+import * as answerStatusDatabaseAdapter from '../../../shared/infrastructure/adapters/answer-status-database-adapter.js';
 import { ActivityAnswer } from '../../domain/models/ActivityAnswer.js';
 
 function _adaptAnswerToDb(answer) {
