+++ api/tests/unit/domain/models/Examiner_test.js
@@ -1,5 +1,5 @@
 import { Answer } from '../../../../src/evaluation/domain/models/Answer.js';
-import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
 import { Examiner } from '../../../../lib/domain/models/Examiner.js';
 import { expect, domainBuilder, sinon } from '../../../test-helper.js';
