+++ api/tests/unit/domain/services/solution-service-qcm_test.js
@@ -1,6 +1,6 @@
 import { expect } from '../../../test-helper.js';
 import * as service from '../../../../lib/domain/services/solution-service-qcm.js';
-import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
 
 describe('Unit | Service | SolutionServiceQCM ', function () {
   describe('if solution type is QCM', function () {
