+++ api/tests/unit/domain/services/solution-service-qrocm-dep_test.js
@@ -1,5 +1,5 @@
 import { expect, catchErr, sinon } from '../../../test-helper.js';
-import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
 import { YamlParsingError } from '../../../../lib/domain/errors.js';
 import {
   getCorrectionDetails,
